Update README.md
I couldn't get a picture of the windows that show up but I included the output of the file in here. If you want to see the windows I talk about the best I can say is to run the code. I made sure there are no errors and it runs.